Break up app in core + ui and inline some scripts
