+++ .eslintrc
@@ -2,6 +2,15 @@
   "extends": "airbnb-base",
   "globals": {
     "__DEV__": false,
-    "__DEMO__": false
+    "Polymer": true
+  },
+  "rules": {
+    "new-cap": 0,
+    "prefer-template": 0,
+    "object-shorthand": 0,
+    "func-names": 0,
+    "prefer-arrow-callback": 0,
+    "no-underscore-dangle": 0,
+    "no-var": 0
   }
 }
