+++ src/cards/ha-card-chooser.js
@@ -4,7 +4,6 @@ import dynamicContentUpdater from '../util/dynamic-content-updater';
 
 require('./ha-camera-card');
 require('./ha-entities-card');
-require('./ha-introduction-card');
 require('./ha-media_player-card');
 
 export default new Polymer({
