--- src/cards/ha-introduction-card.js
@@ -1,18 +1,0 @@
-import Polymer from '../polymer';
-
-require('../components/ha-card');
-
-export default new Polymer({
-  is: 'ha-introduction-card',
-
-  properties: {
-    showInstallInstruction: {
-      type: Boolean,
-      value: __DEMO__,
-    },
-    showHideInstruction: {
-      type: Boolean,
-      value: true,
-    },
-  },
-});
