+++ src/components/ha-card.html
@@ -25,3 +25,23 @@
     <content></content>
   </template>
 </dom-module>
+
+<script>
+Polymer({
+  is: 'ha-card',
+
+  properties: {
+    header: {
+      type: String,
+    },
+    /**
+     * The z-depth of the card, from 0-5.
+     */
+    elevation: {
+      type: Number,
+      value: 1,
+      reflectToAttribute: true,
+    },
+  },
+});
+</script>
