+++ src/components/ha-demo-badge.html
@@ -16,3 +16,9 @@
     ></ha-label-badge>
   </template>
 </dom-module>
+
+<script>
+Polymer({
+  is: 'ha-demo-badge',
+});
+</script>
