--- src/components/ha-demo-badge.js
@@ -1,7 +1,0 @@
-import Polymer from '../polymer';
-
-require('./ha-label-badge');
-
-export default new Polymer({
-  is: 'ha-demo-badge',
-});
