+++ src/components/loading-box.html
@@ -18,3 +18,9 @@
     </div>
   </template>
 </dom-module>
+
+<script>
+Polymer({
+  is: 'loading-box',
+});
+</script>
