--- src/components/loading-box.js
@@ -1,5 +1,0 @@
-import Polymer from '../polymer';
-
-export default new Polymer({
-  is: 'loading-box',
-});
