+++ src/components/logbook-entry.js
@@ -2,7 +2,6 @@ import Polymer from '../polymer';
 
 require('./domain-icon');
 require('./display-time');
-require('./relative-ha-datetime');
 
 export default new Polymer({
   is: 'logbook-entry',
