+++ src/components/state-info.js
@@ -1,7 +1,6 @@
 import Polymer from '../polymer';
 
 require('./entity/state-badge');
-require('./relative-ha-datetime');
 
 export default new Polymer({
   is: 'state-info',
