+++ src/home-assistant.html
@@ -1,14 +1,17 @@
 <link rel='import' href='../bower_components/polymer/polymer.html'>
 <link rel='import' href='../bower_components/paper-spinner/paper-spinner.html'>
+
+<script src='../build/_app_core_compiled.js'></script>
+
 <link rel='import' href='../bower_components/iron-flex-layout/iron-flex-layout-classes.html'>
 <link rel='import' href='./util/roboto.html'>
 <link rel='import' href='../bower_components/paper-styles/typography.html'>
 <link rel="import" href="../bower_components/iron-iconset-svg/iron-iconset-svg.html">
 
+<link rel='import' href='./util/hass-behavior.html'>
 <link rel='import' href='./layouts/login-form.html'>
 <link rel='import' href='./layouts/home-assistant-main.html'>
-
-<link rel='import' href='resources/home-assistant-style.html'>
+<link rel='import' href='./resources/home-assistant-style.html'>
 
 <dom-module id='home-assistant'>
   <style>
@@ -29,9 +32,94 @@
     </template>
 
     <template is='dom-if' if='[[loaded]]'>
-      <home-assistant-main hass='[[hass]]'></home-assistant-main>
+      <home-assistant-main hass='[[hass]]' hidden$='[[!loaded]]'></home-assistant-main>
     </template>
   </template>
 </dom-module>
+
+<script>
+Polymer({
+  is: 'home-assistant',
+
+  hostAttributes: {
+    auth: null,
+    icons: null,
+  },
+
+  behaviors: [window.hassBehavior],
+
+  properties: {
+    hass: {
+      type: Object,
+      value: window.hass,
+    },
+    auth: {
+      type: String,
+    },
+    icons: {
+      type: String,
+    },
+    dataLoaded: {
+      type: Boolean,
+      bindNuclear: function (hass) { return hass.syncGetters.isDataLoaded; },
+    },
+    iconsLoaded: {
+      type: Boolean,
+      value: false,
+    },
+    loaded: {
+      type: Boolean,
+      computed: 'computeLoaded(dataLoaded, iconsLoaded)',
+    },
+  },
+
+  computeLoaded: function (dataLoaded, iconsLoaded) {
+    return dataLoaded && iconsLoaded;
+  },
+
+  computeForceShowLoading: function (dataLoaded, iconsLoaded) {
+    return dataLoaded && !iconsLoaded;
+  },
+
+  loadIcons: function () {
+    // If the import fails, we'll try to import again, must be a server glitch
+    // Since HTML imports only resolve once, we import another url.
+    var success = function () {
+      this.iconsLoaded = true;
+    }.bind(this);
+
+    this.importHref(`/static/mdi-${this.icons}.html`,
+                    success,
+                    function () {
+                      this.importHref('/static/mdi.html', success, success);
+                    });
+  },
+
+  created: function () {
+    if (!('serviceWorker' in navigator)) {
+      return;
+    }
+
+    navigator.serviceWorker.register('/service_worker.js');
+  },
+
+  ready: function () {
+    var hass = this.hass;
+    hass.reactor.batch(function () {
+      // if auth was given, tell the backend
+      if (this.auth) {
+        window.validateAuth(this.hass, this.auth, false);
+      } else if (hass.localStoragePreferences.authToken) {
+        window.validateAuth(this.hass, hass.localStoragePreferences.authToken, true);
+      }
+      hass.navigationActions.showSidebar(hass.localStoragePreferences.showSidebar);
+    });
+
+    hass.startLocalStoragePreferencesSync();
+
+    this.loadIcons();
+  },
+});
+</script>
 
 <script src='../build/_app_compiled.js'></script>
