+++ src/layouts/partial-dev-call-service.html
@@ -55,7 +55,7 @@
 
         <div>
           <div class='header'>Available services</div>
-          <services-list on-service-selected='serviceSelected'></services-list>
+          <services-list on-service-selected='serviceSelected' hass='[[hass]]'></services-list>
         </div>
       </div>
     </partial-base>
