+++ src/layouts/partial-dev-fire-event.html
@@ -49,7 +49,7 @@
 
         <div>
           <div class='header'>Available Events</div>
-          <events-list on-event-selected='eventSelected'></event-list>
+          <events-list on-event-selected='eventSelected' hass='[[hass]]'></event-list>
         </div>
       </div>
     </partial-base>
