+++ src/layouts/partial-dev-set-state.html
@@ -51,7 +51,7 @@
 
         <div>
           <div class='header'>Current entities</div>
-          <entity-list on-entity-selected='entitySelected'></entity-list>
+          <entity-list on-entity-selected='entitySelected' hass='[[hass]]'></entity-list>
         </div>
       </div>
     </partial-base>
