+++ src/layouts/partial-dev-set-state.js
@@ -1,7 +1,6 @@
 import Polymer from '../polymer';
 
 require('./partial-base');
-require('../components/entity-list');
 
 export default new Polymer({
   is: 'partial-dev-set-state',
