+++ src/more-infos/more-info-camera.html
@@ -12,7 +12,38 @@
   </style>
 
   <template>
-    <img class='camera-image' src="[[computeCameraImageUrl(stateObj)]]"
+    <img class='camera-image' src="[[computeCameraImageUrl(hass, stateObj)]]"
          on-load='imageLoaded' />
   </template>
 </dom-module>
+
+<script>
+Polymer({
+  is: 'more-info-camera',
+
+  properties: {
+    hass: {
+      type: Object,
+    },
+
+    stateObj: {
+      type: Object,
+    },
+  },
+
+  imageLoaded: function () {
+    this.fire('iron-resize');
+  },
+
+  computeCameraImageUrl: function (hass, stateObj) {
+    if (hass.demo) {
+      return '/demo/webcam.jpg';
+    } else if (stateObj) {
+      return `/api/camera_proxy_stream/${stateObj.entityId}` +
+             `?token=${stateObj.attributes.access_token}`;
+    }
+    // Return an empty image if no stateObj (= dialog not open)
+    return 'data:image/gif;base64,R0lGODlhAQABAAAAACw=';
+  },
+});
+</script>
