+++ src/more-infos/more-info-configurator.js
@@ -1,7 +1,5 @@
 import Polymer from '../polymer';
 
-require('../components/loading-box');
-
 export default new Polymer({
   is: 'more-info-configurator',
