+++ src/more-infos/more-info-content.js
@@ -11,7 +11,6 @@ require('./more-info-thermostat');
 require('./more-info-script');
 require('./more-info-light');
 require('./more-info-media_player');
-require('./more-info-camera');
 require('./more-info-updater');
 require('./more-info-alarm_control_panel');
 require('./more-info-lock');
