--- src/util/validate-auth.js
@@ -1,6 +1,0 @@
-export default function (hass, authToken, rememberAuth) {
-  hass.authActions.validate(authToken, {
-    rememberAuth,
-    useStreaming: hass.localStoragePreferences.useStreaming,
-  });
-}
